db almost done
