+++ server/db/index.js
@@ -2,11 +2,11 @@
 var db = require('./_db');
 module.exports = db;
 
-var User = require('./models/user');
-var Team = require('./models/Organizations/Team/index');
-var Organization = require('./models/Organizations/Organization/index');
-var UserTeam = require('./models/Organizations/UserTeam/index');
-var UserOrganization = require('./models/Organizations/UserOrganization/index');
+// var User = require('./models/user');
+// var Team = require('./models/Organizations/Team');
+// var Organization = require('./models/Organizations/Organization');
+// var UserTeam = require('./models/Organizations/UserTeam');
+// var UserOrganization = require('./models/Organizations/UserOrganization');
 
-User.belongsToMany(Team, {through: UserTeam});
-User.belongsToMany(Organization, {through: UserOrganization});
+// User.belongsToMany(Team, {through: UserTeam});
+// User.belongsToMany(Organization, {through: UserOrganization});
