+++ server/db/models/user.js
@@ -1,68 +1,68 @@
-'use strict';
-var crypto = require('crypto');
-var _ = require('lodash');
-var Sequelize = require('sequelize');
+// 'use strict';
+// var crypto = require('crypto');
+// var _ = require('lodash');
+// var Sequelize = require('sequelize');
 
-var db = require('../_db');
+// var db = require('../_db');
 
-module.exports = db.define('user', {
-        username: {
-            type: Sequelize.STRING
-        },
-        name: {
-            type: Sequelize.STRING
-        },
-        photo: {
-            type: Sequelize.STRING,
-            validate: {
-                isUrl: true
-            }
-        },
-        email: {
-            unique: true,
-            type: Sequelize.STRING,
-            allowNull: false,
-            validate: {
-                isEmail: true
-            }
-        },
-        password: {
-            type: Sequelize.STRING
-        },
-        salt: {
-            type: Sequelize.STRING
-        },
-        github_id: {
-            type: Sequelize.STRING
-        }
-    }, {
-        instanceMethods: {
-            sanitize: function () {
-                return _.omit(this.toJSON(), ['password', 'salt']);
-            },
-            correctPassword: function (candidatePassword) {
-                return this.Model.encryptPassword(candidatePassword, this.salt) === this.password;
-            }
-        },
-        classMethods: {
-            generateSalt: function () {
-                return crypto.randomBytes(16).toString('base64');
-            },
-            encryptPassword: function (plainText, salt) {
-                var hash = crypto.createHash('sha1');
-                hash.update(plainText);
-                hash.update(salt);
-                return hash.digest('hex');
-            }
-        },
-        hooks: {
-            beforeValidate: function (user) {
-                if (user.changed('password')) {
-                    user.salt = user.Model.generateSalt();
-                    user.password = user.Model.encryptPassword(user.password, user.salt);
-                }
-            }
-        }
-    });
+// module.exports = db.define('user', {
+//         username: {
+//             type: Sequelize.STRING
+//         },
+//         name: {
+//             type: Sequelize.STRING
+//         },
+//         photo: {
+//             type: Sequelize.STRING,
+//             validate: {
+//                 isUrl: true
+//             }
+//         },
+//         email: {
+//             unique: true,
+//             type: Sequelize.STRING,
+//             allowNull: false,
+//             validate: {
+//                 isEmail: true
+//             }
+//         },
+//         password: {
+//             type: Sequelize.STRING
+//         },
+//         salt: {
+//             type: Sequelize.STRING
+//         },
+//         github_id: {
+//             type: Sequelize.STRING
+//         }
+//     }, {
+//         instanceMethods: {
+//             sanitize: function () {
+//                 return _.omit(this.toJSON(), ['password', 'salt']);
+//             },
+//             correctPassword: function (candidatePassword) {
+//                 return this.Model.encryptPassword(candidatePassword, this.salt) === this.password;
+//             }
+//         },
+//         classMethods: {
+//             generateSalt: function () {
+//                 return crypto.randomBytes(16).toString('base64');
+//             },
+//             encryptPassword: function (plainText, salt) {
+//                 var hash = crypto.createHash('sha1');
+//                 hash.update(plainText);
+//                 hash.update(salt);
+//                 return hash.digest('hex');
+//             }
+//         },
+//         hooks: {
+//             beforeValidate: function (user) {
+//                 if (user.changed('password')) {
+//                     user.salt = user.Model.generateSalt();
+//                     user.password = user.Model.encryptPassword(user.password, user.salt);
+//                 }
+//             }
+//         }
+//     });
